Fix test emplacement
